refactor: split analysis/manifest.py into focused modules
